requirements.txt
python-dotenv
openai
unicorn-binance-rest-api
pandas
numpy
rich